fix(check): stabilize nuxt cache ownership
